requirements.txt
asgiref==3.4.1; python_version >= "3.6"
django==3.2.7; python_version >= "3.6"
djangorestframework-simplejwt==5.0.0; python_version >= "3.7"
djangorestframework==3.12.4; python_version >= "3.5"
gunicorn==20.1.0; python_version >= "3.5"
pyjwt==2.3.0; python_version >= "3.7"
pytz==2021.1; python_version >= "3.6"
sqlparse==0.4.2; python_version >= "3.6"
whitenoise==5.3.0; python_version >= "3.5" and python_version < "4"
